Imports, Gradients; Plot Ranges
Removed unnecessary imports; Imporoved plot coloring by using a different color gradient; improved plot ranges, particularly in the "Death Pct" chart ranging from 0 to 0.03 which would be about 99,000 deaths or, for the nation, "worse than the worst average flu season" which is around 70,000. The interpretation of the chart would then be the red areas are contributing towards that, the green areas doing much better.